Implement support for user specified xml encoders.
This is a combination of 10 commits.
Refactor XMLRPCEncoder to XMLRPCEncoderImpl.

This change renames XMLRPCEncoder to XMLRPCEncoderImpl to imply that
XMLRPCEncoder is actually an implentation of a interface yet to be defined. The
yet to be defined interface will allow users of the xmlrpc to prvide theier own
encoders to support XMLRPC extensions.

Create a protocol for XMLRPCEncoderImpl to implement.

This protocol will be used instead of XMLRPCEncoderImpl directly.

Have XMLRPC Request use the new XMLRPCEncoder protocol instead of XMLRPCEncoderImpl

Add a setEncoder method.

Remove header created by XCode

Refactor XMLRPCDencoderImpl to XMLRPCDefaultEndcoder

This refactor was made in response to a discussion on naming conventions here:
https://github.com/eczarny/xmlrpc/pull/17#r645514

Remove unnecessary prototypes from XMLRPCDefaultEncoder

The methods in defined in XMLRPCDefaultEncoder.h are already defined in the
XMLRPCEncoder protocol and there are are redundant.

Add a constuctor to initialize XMLRPCRequest with a given encoder.

HAve XMLRPCREquest take ownership of encoders assigned to it.

This commit causes XMLRPCRequest to retain all XMLRPCEncoders assigned to it.

Call default constructor when initailizing the default XMLRPC Encoder