Include Connection In Request Object
Adds a pointer to the xmlrpc connection inside of the request itself,
this allows for the delegate response handler to access the connection
via

[request connection];

and more importantly the connection's uuid identifier via

[[request connection] identifier];

this allows for a single delegate class to dispatch responses uniquely
based on their uuid, essentially providing a direct mapping external to
the rpc client.

This patch also should not impact existing clients as it doesn't change
any delegate requirements, but simply passively exposes the connection
to those who desire it.