Added -description Method with ARC Support
I thought it would be nice to have a -description method on the Response so
that debugging via logging was a lot simpler. This turns out to be very
helpful. At the same time, I added in the ARC support so it will work with
retain/release as well as ARC.