Added ARC Support
I have added in the checks for Automatic Reference Counting by the LLVM
compiler. This is just a compile-time option for the code, so that it can be
used in retain/release schemes as well as ARC schemes. It's a pretty simple
matter of removing the unnecessary calls that the compiler adds in anyway.